add tab reordering and a close button
drag tab to re-order it, but the +NEW (first tab) cannot be moved and
tabs cannot be dragged befor it - any attemp will put the move tab
just after it.

The inactive screen is now two buttons the to half is start/restart
(turns green on hover); and the bottom half is close (turns red on
hover).

Signed-off-by: Christopher Hall <hsw@ms2.hinet.net>