save and restore session, new configuration file location
The old .conlectermrc is renamed and moved to the configuration directory
i.e. the default is:  ${HOME}/.conlecterm/config.rc

The list of tabs/buttons is saved in the configuration directory
under the name:  session-<SESSION-NAME>.rc
e.g. the session default is:  session-default.rc

The --config option now specifies the path to the configuration
directory wher the confic.rc and session-*.rc are kept

Signed-off-by: Christopher Hall <hsw@ms2.hinet.net>